+++ R/tidy_beeca.R
@@ -28,7 +28,8 @@
 #'        }
 #' @export
 #' @examples
-#' # example use of tidy_beeca
+#' library(broom)
+#'
 #' # Set treatment to a factor
 #' trial01$trtp <- factor(trial01$trtp)
 #'
@@ -36,9 +37,10 @@
 #' fit1 <- glm(aval ~ trtp + bl_cov, family="binomial", data=trial01) |>
 #'   get_marginal_effect(trt="trtp", method="Ye", contrast="diff")
 #'
-#' # compare with broom tidy. one works on logistic model(working model)
-#' # tidy_beeca tidies the analyses from g-computation
+#' # compare with broom tidy. tidy(fit) works on logistic model(working model)
 #' broom::tidy(fit1)
+#'
+#' # tidy_beeca tidies the analyses from g-computation
 #' tidy_beeca(fit1)
 #'
 #' # with confidence intervals
@@ -46,8 +48,6 @@
 #' tidy_beeca(fit1, conf.int = TRUE)
 #'
 tidy_beeca <- function(x, conf.int = FALSE, conf.level = 0.95) {
-
-  results <- NULL
 
   # check inputs ---------------------------------------------------------------